+++ src/tz/db/zoneinfo/enabled.rs
@@ -8,7 +8,10 @@ use std::{
     fs::File,
     io::Read,
     path::{Path, PathBuf},
-    sync::{Arc, RwLock},
+    sync::{
+        atomic::{AtomicUsize, Ordering},
+        Arc, RwLock,
+    },
     time::Duration,
 };
 
@@ -256,17 +259,27 @@ impl CachedTimeZone {
         info: &ZoneInfoName,
         ttl: Duration,
     ) -> Result<CachedTimeZone, Error> {
-        let path = &info.inner.full;
-        let mut file =
-            File::open(path).map_err(|e| Error::io(e).path(path))?;
-        let mut data = vec![];
-        file.read_to_end(&mut data).map_err(|e| Error::io(e).path(path))?;
-        let tz = TimeZone::tzif(&info.inner.original, &data)
-            .map_err(|e| e.path(path))?;
-        let name = info.clone();
-        let last_modified = util::fs::last_modified_from_file(path, &file);
-        let expiration = Expiration::after(ttl);
-        Ok(CachedTimeZone { tz, name, expiration, last_modified })
+        fn imp(
+            info: &ZoneInfoName,
+            ttl: Duration,
+        ) -> Result<CachedTimeZone, Error> {
+            let path = info.path();
+            let mut file =
+                File::open(path).map_err(|e| Error::io(e).path(path))?;
+            let mut data = vec![];
+            file.read_to_end(&mut data)
+                .map_err(|e| Error::io(e).path(path))?;
+            let tz = TimeZone::tzif(&info.inner.original, &data)
+                .map_err(|e| e.path(path))?;
+            let name = info.clone();
+            let last_modified = util::fs::last_modified_from_file(path, &file);
+            let expiration = Expiration::after(ttl);
+            Ok(CachedTimeZone { tz, name, expiration, last_modified })
+        }
+
+        let result = imp(info, ttl);
+        info.set_validity(result.is_ok());
+        result
     }
 
     /// Returns true if this time zone has gone stale and should, at minimum,
@@ -299,7 +312,7 @@ impl CachedTimeZone {
             return false;
         };
         let Some(new_last_modified) =
-            util::fs::last_modified_from_path(&info.inner.full)
+            util::fs::last_modified_from_path(info.path())
         else {
             trace!(
                 "revalidation for {} failed because new last modified time \
@@ -448,7 +461,11 @@ impl ZoneInfoNamesInner {
 
     /// Returns all available time zone names.
     fn available(&self) -> Vec<String> {
-        self.names.iter().map(|n| n.inner.original.clone()).collect()
+        self.names
+            .iter()
+            .filter(|n| n.is_valid())
+            .map(|n| n.inner.original.clone())
+            .collect()
     }
 
     /// Attempts a refresh, but only follows through if the TTL has been
@@ -502,7 +519,7 @@ struct ZoneInfoName {
     inner: Arc<ZoneInfoNameInner>,
 }
 
-#[derive(Clone, Debug)]
+#[derive(Debug)]
 struct ZoneInfoNameInner {
     /// A file path resolvable to the corresponding file relative to the
     /// working directory of this program.
@@ -516,6 +533,13 @@ struct ZoneInfoNameInner {
     /// The lowercase version of `original`. This is how we determine name
     /// equality.
     lower: String,
+    /// The known validity state of this time zone name. `0` means unknown (and
+    /// thus we need to check), `1` means "presumably valid" and `2` means
+    /// "known invalid." The "presumably valid" means that the file has a
+    /// 4-byte TZif header and the odds of a false positive a low enough that
+    /// we can and should behave as if that file is actually TZif and thus a
+    /// valid IANA time zone identifier.
+    validity: AtomicUsize,
 }
 
 impl ZoneInfoName {
@@ -528,9 +552,18 @@ impl ZoneInfoName {
         let original = parse::os_str_utf8(time_zone_name.as_os_str())
             .map_err(|err| err.path(base))?;
         let lower = original.to_ascii_lowercase();
-        let inner =
-            ZoneInfoNameInner { full, original: original.to_string(), lower };
+        let inner = ZoneInfoNameInner {
+            full,
+            original: original.to_string(),
+            lower,
+            validity: AtomicUsize::new(ZONE_INFO_NAME_UNKNOWN),
+        };
         Ok(ZoneInfoName { inner: Arc::new(inner) })
+    }
+
+    /// Returns the path to the corresponding (presumed) TZif file.
+    fn path(&self) -> &Path {
+        &self.inner.full
     }
 
     /// Returns the original name of this time zone.
@@ -541,6 +574,78 @@ impl ZoneInfoName {
     /// Returns the lowercase name of this time zone.
     fn lower(&self) -> &str {
         &self.inner.lower
+    }
+
+    /// Returns true if it is presumed that this points to a valid TZif file.
+    ///
+    /// The result of this function may use a cached value.
+    fn is_valid(&self) -> bool {
+        let validity = self.inner.validity.load(Ordering::Relaxed);
+        if validity == ZONE_INFO_NAME_VALID {
+            return true;
+        } else if validity == ZONE_INFO_NAME_INVALID {
+            return false;
+        }
+        if self.is_valid_impl() {
+            self.inner.validity.store(ZONE_INFO_NAME_VALID, Ordering::Relaxed);
+            true
+        } else {
+            self.inner
+                .validity
+                .store(ZONE_INFO_NAME_INVALID, Ordering::Relaxed);
+            false
+        }
+    }
+
+    /// Marks this zone info name as known to be valid or invalid.
+    ///
+    /// e.g., After doing a successful time zone lookup.
+    fn set_validity(&self, is_valid: bool) {
+        let validity = if is_valid {
+            ZONE_INFO_NAME_VALID
+        } else {
+            ZONE_INFO_NAME_INVALID
+        };
+        self.inner.validity.store(validity, Ordering::Relaxed);
+    }
+
+    /// Performs the actual validity check.
+    ///
+    /// This is generally only needed for APIs like
+    /// `TimeZoneDatabase::available()`, where we don't want to load every time
+    /// zone into memory but we also don't want to return IANA time zone ids
+    /// like `zone.tab`. (Because a `/usr/share/zoneinfo` directory might have
+    /// random junk in it.)
+    fn is_valid_impl(&self) -> bool {
+        let path = self.path();
+        let mut f = match File::open(path) {
+            Ok(f) => f,
+            Err(_err) => {
+                trace!("failed to open {}: {_err}", path.display());
+                return false;
+            }
+        };
+        let mut buf = [0; 4];
+        if let Err(_err) = f.read_exact(&mut buf) {
+            trace!(
+                "failed to read first 4 bytes of {}: {_err}",
+                path.display()
+            );
+            return false;
+        }
+        if !is_possibly_tzif(&buf) {
+            // This is a trace because it's perfectly normal for a
+            // non-TZif file to be in a zoneinfo directory. But it could
+            // still be potentially useful debugging info.
+            trace!(
+                "found file {} that isn't TZif since its first \
+                 four bytes are {:?}",
+                path.display(),
+                crate::util::escape::Bytes(&buf),
+            );
+            return false;
+        }
+        true
     }
 }
 
@@ -570,6 +675,10 @@ impl core::hash::Hash for ZoneInfoName {
     }
 }
 
+static ZONE_INFO_NAME_UNKNOWN: usize = 0;
+static ZONE_INFO_NAME_VALID: usize = 1;
+static ZONE_INFO_NAME_INVALID: usize = 2;
+
 /// Recursively walks the given directory and returns the names of all time
 /// zones found.
 ///
@@ -581,6 +690,16 @@ impl core::hash::Hash for ZoneInfoName {
 ///
 /// The names returned are sorted in lexicographic order according to the
 /// lowercase form of each name.
+///
+/// # Performance
+///
+/// Note that this routine is written in a way that, at least on Unix, we
+/// should not be doing a syscall for every file. We need to do one for every
+/// directory, but that should be comparatively rare. It's done this way to
+/// avoid long initialization times when `/usr/share/zoneinfo` is on a slow
+/// file system.
+///
+/// See: https://github.com/BurntSushi/jiff/issues/366
 fn walk(start: &Path) -> Result<Vec<ZoneInfoName>, Error> {
     let mut first_err: Option<Error> = None;
     let mut seterr = |path: &Path, err: Error| {
@@ -635,40 +754,7 @@ fn walk(start: &Path) -> Result<Vec<ZoneInfoName>, Error> {
             // We assume symlinks are files, although this may not be
             // appropriate. If we need to also handle the case when they're
             // directories, then we'll need to add symlink loop detection.
-            //
-            // Otherwise, at this point, we peek at the first few bytes of a
-            // file to do a low false positive and never false negative check
-            // for a TZif file.
 
-            let mut f = match File::open(&path) {
-                Ok(f) => f,
-                Err(err) => {
-                    trace!("failed to open {}: {err}", path.display());
-                    seterr(&path, Error::io(err));
-                    continue;
-                }
-            };
-            let mut buf = [0; 4];
-            if let Err(err) = f.read_exact(&mut buf) {
-                trace!(
-                    "failed to read first 4 bytes of {}: {err}",
-                    path.display()
-                );
-                seterr(&path, Error::io(err));
-                continue;
-            }
-            if !is_possibly_tzif(&buf) {
-                // This is a trace because it's perfectly normal for a
-                // non-TZif file to be in a zoneinfo directory. But it could
-                // still be potentially useful debugging info.
-                trace!(
-                    "found file {} that isn't TZif since its first \
-                     four bytes are {:?}",
-                    path.display(),
-                    crate::util::escape::Bytes(&buf),
-                );
-                continue;
-            }
             let time_zone_name = match path.strip_prefix(start) {
                 Ok(time_zone_name) => time_zone_name,
                 Err(err) => {